+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Bug Fix: [Validate field names are unique to an object/interface/input object](https://github.com/absinthe-graphql/absinthe/pull/1135)
+- Bug Fix: [Validate variable usage in according to spec](https://github.com/absinthe-graphql/absinthe/pull/1141)
 
 ## 1.6.6